refactor(app) update simulator dates logic
